farmwise/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-genai>=0.5.0
python-dotenv==1.0.0